+++ dpctl-capi/tests/test_sycl_program_interface.cpp
@@ -1,4 +1,4 @@
-//===---------- test_sycl_program_interface.cpp - dpctl-C_API --*-- C++ -*-===//
+//===---------- test_sycl_program_interface.cpp - dpctl-C_API ---*- C++ -*-===//
 //
 //               Data Parallel Control Library (dpCtl)
 //
@@ -29,7 +29,7 @@
 #include "dpctl_sycl_program_interface.h"
 #include "dpctl_sycl_queue_interface.h"
 #include "dpctl_sycl_queue_manager.h"
-
+#include "Config/dpctl_config.h"
 #include <array>
 #include <fstream>
 #include <filesystem>
@@ -152,7 +152,7 @@ TEST_F (TestDPCTLSyclProgramInterface, CheckCreateFromOCLSource)
     auto QueueRef = DPCTLQueueMgr_GetQueue(DPCTL_OPENCL, DPCTL_GPU, 0);
     auto CtxRef = DPCTLQueue_GetContext(QueueRef);
     auto PRef = DPCTLProgram_CreateFromOCLSource(CtxRef, CLProgramStr,
-                                                CompileOpts);
+                                                 CompileOpts);
     ASSERT_TRUE(PRef != nullptr);
     ASSERT_TRUE(DPCTLProgram_HasKernel(PRef, "add"));
     ASSERT_TRUE(DPCTLProgram_HasKernel(PRef, "axpy"));
@@ -162,15 +162,16 @@ TEST_F (TestDPCTLSyclProgramInterface, CheckCreateFromOCLSource)
     DPCTLProgram_Delete(PRef);
 }
 
-TEST_F (TestDPCTLSyclProgramInterface, CheckCreateFromOCLSpirv)
+TEST_F (TestDPCTLSyclProgramInterface, CheckCreateFromSpirvOCL)
 {
     if(!nOpenCLGpuQ)
         GTEST_SKIP_("Skipping as no OpenCL GPU device found.\n");
 
     auto QueueRef = DPCTLQueueMgr_GetQueue(DPCTL_OPENCL, DPCTL_GPU, 0);
     auto CtxRef = DPCTLQueue_GetContext(QueueRef);
-    auto PRef = DPCTLProgram_CreateFromOCLSpirv(CtxRef, spirvBuffer.data(),
-                                               spirvFileSize);
+    auto PRef = DPCTLProgram_CreateFromSpirv(CtxRef, spirvBuffer.data(),
+                                             spirvFileSize,
+                                             nullptr);
     ASSERT_TRUE(PRef != nullptr);
     ASSERT_TRUE(DPCTLProgram_HasKernel(PRef, "add"));
     ASSERT_TRUE(DPCTLProgram_HasKernel(PRef, "axpy"));
@@ -179,6 +180,28 @@ TEST_F (TestDPCTLSyclProgramInterface, CheckCreateFromOCLSpirv)
     DPCTLContext_Delete(CtxRef);
     DPCTLProgram_Delete(PRef);
 }
+
+#ifdef DPCTL_ENABLE_LO_PROGRAM_CREATION
+TEST_F (TestDPCTLSyclProgramInterface, CheckCreateFromSpirvL0)
+{
+    auto nL0GpuQ = DPCTLQueueMgr_GetNumQueues(DPCTL_LEVEL_ZERO, DPCTL_GPU);
+    if(!nL0GpuQ)
+        GTEST_SKIP_("Skipping as no OpenCL GPU device found.\n");
+
+    auto QueueRef = DPCTLQueueMgr_GetQueue(DPCTL_LEVEL_ZERO, DPCTL_GPU, 0);
+    auto CtxRef = DPCTLQueue_GetContext(QueueRef);
+    auto PRef = DPCTLProgram_CreateFromSpirv(CtxRef, spirvBuffer.data(),
+                                             spirvFileSize,
+                                             nullptr);
+    ASSERT_TRUE(PRef != nullptr);
+    ASSERT_TRUE(DPCTLProgram_HasKernel(PRef, "add"));
+    ASSERT_TRUE(DPCTLProgram_HasKernel(PRef, "axpy"));
+
+    DPCTLQueue_Delete(QueueRef);
+    DPCTLContext_Delete(CtxRef);
+    DPCTLProgram_Delete(PRef);
+}
+#endif
 
 TEST_F (TestDPCTLSyclProgramInterface, CheckGetKernelOCLSource)
 {
@@ -207,15 +230,15 @@ TEST_F (TestDPCTLSyclProgramInterface, CheckGetKernelOCLSource)
     DPCTLProgram_Delete(PRef);
 }
 
-TEST_F (TestDPCTLSyclProgramInterface, CheckGetKernelOCLSpirv)
+TEST_F (TestDPCTLSyclProgramInterface, CheckGetKernelSpirv)
 {
     if(!nOpenCLGpuQ)
         GTEST_SKIP_("Skipping as no OpenCL GPU device found.\n");
 
     auto QueueRef = DPCTLQueueMgr_GetQueue(DPCTL_OPENCL, DPCTL_GPU, 0);
     auto CtxRef = DPCTLQueue_GetContext(QueueRef);
-    auto PRef = DPCTLProgram_CreateFromOCLSpirv(CtxRef, spirvBuffer.data(),
-                                               spirvFileSize);
+    auto PRef = DPCTLProgram_CreateFromSpirv(CtxRef, spirvBuffer.data(),
+                                             spirvFileSize, nullptr);
     auto AddKernel = DPCTLProgram_GetKernel(PRef, "add");
     auto AxpyKernel = DPCTLProgram_GetKernel(PRef, "axpy");