+++ dpctl/tests/test_sycl_program.py
@@ -87,9 +87,16 @@ class TestProgramFromSPRIV(unittest.TestCase):
     "No Level0 GPU queues available",
 )
 class TestProgramForLevel0GPU(unittest.TestCase):
-    @unittest.expectedFailure
-    def test_create_program_from_spirv(self):
 
+    import sys
+
+    # Level zero program creation from a SPIR-V binary is not supported
+    # on Windows.
+    @unittest.skipIf(
+        sys.platform in ["win32", "cygwin"],
+        "Level Zero module creation unsupported on Windows.",
+    )
+    def test_create_program_from_spirv(self):
         CURR_DIR = os.path.dirname(os.path.abspath(__file__))
         spirv_file = os.path.join(CURR_DIR, "input_files/multi_kernel.spv")
         with open(spirv_file, "rb") as fin: